Fixed exception, when SCP leaves, but noone can switch it

+++ betterRP/Handlers/PlayerLeave.cs
@@ -20,7 +20,7 @@ namespace betterRP
         Random rnd = new Random();
         public void OnDisconnect(LeftEventArgs ev)
         {
-            if (ev.Player.IsScp)
+            if (ev.Player.IsScp && Player.List.Where(p => p.IsDead).Count() > 0)
             {
                 var SCP = ev.Player.Role;
                 Log.Info(SCP.ToString());
@@ -31,7 +31,7 @@ namespace betterRP
                 Players[randomplayer].Role = SCP;
                 Players[randomplayer].Position = Position;
             }
-            else
+            else if(!ev.Player.IsScp)
             {
                 ev.Player.Kill(DamageTypes.Scp049);
             }
